Fixes after review. Javadocs added, default constructors removed, introduced immutable fields and classes
