+++ BattleSheepLib/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/models/QueueObject.java
@@ -2,5 +2,8 @@ package pl.jakubowskiprzemyslaw.tajgertim.models;
 
 import java.io.Serializable;
 
+/**
+ * Marking interface for objects which will be send to queues
+ */
 public interface QueueObject extends Serializable {
 }
