+++ BattleSheepLib/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/models/coordinates/Coordinate.java
@@ -4,16 +4,12 @@ import java.io.Serializable;
 
 public class Coordinate implements Serializable {
 
-    private int x;
-    private int y;
+    private final int x;
+    private final int y;
 
     public Coordinate(int x, int y) {
         this.x = x;
         this.y = y;
-    }
-
-    public Coordinate() {
-        this(0,0);
     }
 
     @Override
@@ -28,15 +24,8 @@ public class Coordinate implements Serializable {
         return x;
     }
 
-    public void setX(int x) {
-        this.x = x;
-    }
-
     public int getY() {
         return y;
     }
 
-    public void setY(int y) {
-        this.y = y;
-    }
 }
