+++ BattleSheepLib/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/models/playeraction/action/Action.java
@@ -2,5 +2,8 @@ package pl.jakubowskiprzemyslaw.tajgertim.models.playeraction.action;
 
 import java.io.Serializable;
 
+/**
+ * Marking interface for actions which Player can perform
+ */
 public interface Action extends Serializable {
 }
