+++ BattleSheepLib/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/models/round/NextRoundStatus.java
@@ -2,6 +2,12 @@ package pl.jakubowskiprzemyslaw.tajgertim.models.round;
 
 import pl.jakubowskiprzemyslaw.tajgertim.models.QueueObject;
 
+/**
+ * Class used for communication between microservices Judge and PlayingStateMachine,
+ * to inform PlayingStateMachine which player should
+ * play next or if game should end
+ */
+
 public class NextRoundStatus implements QueueObject {
     RoundStatus roundStatus;
